+++ src/test/java/racingcar/RaceControllerTest.java
@@ -1,8 +1,8 @@
-package step3;
+package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.RaceController;
+import racingcar.contorller.RaceController;
 
 import java.util.List;
